Change method for plugin import. Split off more code to separate modules.
