+++ pyomo/contrib/gdpopt/GDPopt.py
@@ -28,36 +28,22 @@ import logging
 import pyomo.common.plugin
 from pyomo.common.config import (ConfigBlock, ConfigList, ConfigValue, In,
                                  NonNegativeFloat, NonNegativeInt)
-from pyomo.contrib.gdpopt.cut_generation import (add_integer_cut,
-                                                 add_outer_approximation_cuts)
-from pyomo.contrib.gdpopt.loa import solve_LOA_subproblem, solve_LOA_master
-from pyomo.contrib.gdpopt.gloa import solve_GLOA_master, solve_global_NLP
-from pyomo.contrib.gdpopt.master_initialize import (init_custom_disjuncts,
-                                                    init_max_binaries,
-                                                    init_set_covering,
-                                                    init_fixed_disjuncts)
-from pyomo.contrib.gdpopt.util import (GDPoptSolveData, _DoNothing, a_logger,
-                                       algorithm_should_terminate,
+from pyomo.contrib.gdpopt.data_class import GDPoptSolveData
+from pyomo.contrib.gdpopt.iterate import GDPopt_iteration_loop
+from pyomo.contrib.gdpopt.master_initialize import (GDPopt_initialize_master,
+                                                    valid_init_strategies)
+from pyomo.contrib.gdpopt.util import (_DoNothing, a_logger,
                                        build_ordered_component_lists,
                                        clone_orig_model_with_lists,
                                        copy_var_list_values, model_is_valid,
                                        record_original_model_statistics,
                                        record_working_model_statistics,
                                        reformulate_integer_variables)
-from pyomo.core.base import Block, ConstraintList, Suffix, value
+from pyomo.core.base import Block, ConstraintList, value
 from pyomo.core.kernel import ComponentMap
 from pyomo.opt.base import IOptSolver
 
 __version__ = (0, 4, 0)
-
-# Valid initialization strategies
-valid_init_strategies = {
-    'no_init': _DoNothing,
-    'set_covering': init_set_covering,
-    'max_binary': init_max_binaries,
-    'fix_disjuncts': init_fixed_disjuncts,
-    'custom_disjuncts': init_custom_disjuncts
-}
 
 
 class GDPoptSolver(pyomo.common.plugin.Plugin):
@@ -293,10 +279,10 @@ class GDPoptSolver(pyomo.common.plugin.Plugin):
             solve_data.feasible_solution_improved = False
 
             # Initialize the master problem
-            self._GDPopt_initialize_master(solve_data, config)
+            GDPopt_initialize_master(solve_data, config)
 
             # Algorithm main loop
-            self._GDPopt_iteration_loop(solve_data, config)
+            GDPopt_iteration_loop(solve_data, config)
 
             # Update values in working model
             copy_var_list_values(
@@ -319,78 +305,3 @@ class GDPoptSolver(pyomo.common.plugin.Plugin):
             config.logger.setLevel(old_logger_level)
             if created_GDPopt_block:
                 model.del_component('GDPopt_utils')
-
-    def _GDPopt_initialize_master(self, solve_data, config):
-        """Initialize the decomposition algorithm.
-
-        This includes generating the initial cuts require to build the master
-        problem.
-
-        """
-        config.logger.info("---Starting GDPopt initialization---")
-        m = solve_data.working_model
-        if not hasattr(m, 'dual'):  # Set up dual value reporting
-            m.dual = Suffix(direction=Suffix.IMPORT)
-        m.dual.activate()
-
-        # Set up the linear GDP model
-        solve_data.linear_GDP = m.clone()
-        # deactivate nonlinear constraints
-        for c in solve_data.linear_GDP.GDPopt_utils.\
-                working_nonlinear_constraints:
-            c.deactivate()
-
-        # Initialization strategies
-        init_strategy = valid_init_strategies.get(config.init_strategy, None)
-        if init_strategy is not None:
-            init_strategy(solve_data, config)
-        else:
-            raise ValueError(
-                'Unknown initialization strategy: %s. '
-                'Valid strategies include: %s'
-                % (config.init_strategy,
-                   ", ".join(k for (k, v) in valid_init_strategies.items()
-                             if v is not None)))
-
-    def _GDPopt_iteration_loop(self, solve_data, config):
-        """Algorithm main loop.
-
-        Returns True if successful convergence is obtained. False otherwise.
-
-        """
-        while solve_data.master_iteration < config.iterlim:
-            # print line for visual display
-            solve_data.master_iteration += 1
-            solve_data.mip_iteration = 0
-            solve_data.nlp_iteration = 0
-            config.logger.info(
-                '---GDPopt Master Iteration %s---'
-                % solve_data.master_iteration)
-            # solve MILP master problem
-            if solve_data.current_strategy == 'LOA':
-                mip_results = solve_LOA_master(solve_data, config)
-            elif solve_data.current_strategy == 'GLOA':
-                mip_results = solve_GLOA_master(solve_data, config)
-            if mip_results:
-                _, mip_var_values = mip_results
-            # Check termination conditions
-            if algorithm_should_terminate(solve_data, config):
-                break
-            # Solve NLP subproblem
-            if solve_data.current_strategy == 'LOA':
-                nlp_result = solve_LOA_subproblem(
-                    mip_var_values, solve_data, config)
-                nlp_feasible, nlp_var_values, nlp_duals = nlp_result
-                if nlp_feasible:
-                    add_outer_approximation_cuts(
-                        nlp_var_values, nlp_duals, solve_data, config)
-            elif solve_data.current_strategy == 'GLOA':
-                nlp_result = solve_global_NLP(
-                    mip_var_values, solve_data, config)
-                # TODO add affine cuts
-            add_integer_cut(
-                mip_var_values, solve_data, config, feasible=nlp_feasible)
-
-            # Check termination conditions
-            if algorithm_should_terminate(solve_data, config):
-                break
