--- pyomo/contrib/gdpopt/__init__.py
@@ -1,1 +1,0 @@
-import pyomo.contrib.gdpopt.GDPopt
