+++ pyomo/contrib/gdpopt/data_class.py
@@ -1,0 +1,18 @@
+"""Module for GDPopt data classes."""
+
+
+class GDPoptSolveData(object):
+    """Data container to hold solve-instance data.
+
+    Key attributes:
+        - original_model: the original model that the user gave us to solve
+        - working_model: the original model after preprocessing
+        - linear_GDP: the linear-discrete master problem
+
+    """
+    pass
+
+
+class MasterProblemResult(object):
+    """Data class for master problem results data."""
+    pass
