+++ pyomo/contrib/gdpopt/iterate.py
@@ -1,0 +1,122 @@
+"""Iteration code."""
+from __future__ import division
+
+from pyomo.contrib.gdpopt.cut_generation import (add_integer_cut,
+                                                 add_outer_approximation_cuts)
+from pyomo.contrib.gdpopt.gloa import solve_GLOA_master, solve_global_NLP
+from pyomo.contrib.gdpopt.loa import solve_LOA_master, solve_LOA_subproblem
+
+
+def GDPopt_iteration_loop(solve_data, config):
+    """Algorithm main loop.
+
+    Returns True if successful convergence is obtained. False otherwise.
+
+    """
+    while solve_data.master_iteration < config.iterlim:
+        # Set iteration counters for new master iteration.
+        solve_data.master_iteration += 1
+        solve_data.mip_iteration = 0
+        solve_data.nlp_iteration = 0
+
+        # print line for visual display
+        config.logger.info(
+            '---GDPopt Master Iteration %s---'
+            % solve_data.master_iteration)
+
+        # solve linear master problem
+        if solve_data.current_strategy == 'LOA':
+            mip_results = solve_LOA_master(solve_data, config)
+        elif solve_data.current_strategy == 'GLOA':
+            mip_results = solve_GLOA_master(solve_data, config)
+        if mip_results:
+            _, mip_var_values = mip_results
+        # Check termination conditions
+        if algorithm_should_terminate(solve_data, config):
+            break
+        # Solve NLP subproblem
+        if solve_data.current_strategy == 'LOA':
+            nlp_result = solve_LOA_subproblem(
+                mip_var_values, solve_data, config)
+            nlp_feasible, nlp_var_values, nlp_duals = nlp_result
+            if nlp_feasible:
+                add_outer_approximation_cuts(
+                    nlp_var_values, nlp_duals, solve_data, config)
+        elif solve_data.current_strategy == 'GLOA':
+            nlp_result = solve_global_NLP(
+                mip_var_values, solve_data, config)
+            # TODO add affine cuts
+        add_integer_cut(
+            mip_var_values, solve_data, config, feasible=nlp_feasible)
+
+        # Check termination conditions
+        if algorithm_should_terminate(solve_data, config):
+            break
+
+
+def algorithm_should_terminate(solve_data, config):
+    """Check if the algorithm should terminate.
+
+    Termination conditions based on solver options and progress.
+
+    """
+    # Check bound convergence
+    if solve_data.LB + config.bound_tolerance >= solve_data.UB:
+        config.logger.info(
+            'GDPopt exiting on bound convergence. '
+            'LB: %s + (tol %s) >= UB: %s' %
+            (solve_data.LB, config.bound_tolerance,
+             solve_data.UB))
+        return True
+
+    # Check iteration limit
+    if solve_data.master_iteration >= config.iterlim:
+        config.logger.info(
+            'GDPopt unable to converge bounds '
+            'after %s master iterations.'
+            % (solve_data.master_iteration,))
+        config.logger.info(
+            'Final bound values: LB: %s  UB: %s'
+            % (solve_data.LB, solve_data.UB))
+        return True
+
+    if not algorithm_is_making_progress(solve_data, config):
+        config.logger.debug(
+            'Algorithm is not making enough progress. '
+            'Exiting iteration loop.')
+        return True
+    return False
+
+
+def algorithm_is_making_progress(solve_data, config):
+    """Make sure that the algorithm is making sufficient progress
+    at each iteration to continue."""
+
+    # TODO if backtracking is turned on, and algorithm visits the same point
+    # twice without improvement in objective value, turn off backtracking.
+
+    # TODO stop iterations if feasible solutions not progressing for a number
+    # of iterations.
+
+    # If the hybrid algorithm is not making progress, switch to OA.
+    # required_feas_prog = 1E-6
+    # if solve_data.working_model.GDPopt_utils.objective.sense == minimize:
+    #     sign_adjust = 1
+    # else:
+    #     sign_adjust = -1
+
+    # Maximum number of iterations in which feasible bound does not
+    # improve before terminating algorithm
+    # if (len(feas_prog_log) > config.algorithm_stall_after and
+    #     (sign_adjust * (feas_prog_log[-1] + required_feas_prog)
+    #      >= sign_adjust *
+    #      feas_prog_log[-1 - config.algorithm_stall_after])):
+    #     config.logger.info(
+    #         'Feasible solutions not making enough progress '
+    #         'for %s iterations. Algorithm stalled. Exiting.\n'
+    #         'To continue, increase value of parameter '
+    #         'algorithm_stall_after.'
+    #         % (config.algorithm_stall_after,))
+    #     return False
+
+    return True
