+++ pyomo/contrib/gdpopt/plugins.py
@@ -1,0 +1,2 @@
+def load():
+    import pyomo.contrib.gdpopt.GDPopt
